+++ src/script/components/MessagesList/Message/ContentMessage/MessageActions/MessageActions.styles.ts
@@ -94,3 +94,7 @@ export const getActionsMenuCSS = (isActive?: boolean): CSSObject => {
 export const messageWithHeaderTop: CSSObject = {
   top: '2.8rem',
 };
+
+export const reactionImgSize = {
+  width: '1rem',
+};
